Fix a query test

+++ src/main/scala/name/amadoucisse/restoo/domain/entries/EntryRepositoryAlgebra.scala
@@ -7,5 +7,5 @@ import items.ItemId
 trait EntryRepositoryAlgebra[F[_]] {
   def create(entry: Entry): F[Entry]
 
-  def count(id: ItemId): F[Option[Int]]
+  def count(id: ItemId): F[Option[Long]]
 }
